requirements.txt
fastapi==0.111.*
uvicorn[standard]==0.30.*
httpx==0.27.*
pydantic-settings==2.4.*
cachetools==5.3.*
slowapi==0.1.*
python-dotenv==1.0.*
structlog==24.*
pytest==8.*
